podcast_process/requirements.txt
aio-pika
requests
boto3
python-box
click
mutagen
protobuf